refactor(handleFiles): added supported file extension checking in class CheckSupportedFileExtension in new handleFiles package.
