--- core/src/main/java/dev/arnoldatse/opensource/look4dev/core/fileStorage/adapters/FileStoragePersistAdapter.java
@@ -1,9 +1,0 @@
-package dev.arnoldatse.opensource.look4dev.core.fileStorage.adapters;
-
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.FailedToStoreFileException;
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.enums.FilesDirectories;
-
-public interface FileStoragePersistAdapter<T> {
-    void store(T file, String fileName, String fileExtension, FilesDirectories dir) throws FailedToStoreFileException;
-    void delete(String fileName, FilesDirectories dir);
-}
