+++ core/src/main/java/dev/arnoldatse/opensource/look4dev/core/handleFiles/fileStorage/FailedToStoreFileException.java
@@ -1,4 +1,4 @@
-package dev.arnoldatse.opensource.look4dev.core.fileStorage;
+package dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage;
 
 public class FailedToStoreFileException extends Exception{
     public FailedToStoreFileException(String msg){
