+++ core/src/main/java/dev/arnoldatse/opensource/look4dev/core/handleFiles/fileStorage/adapters/FileStorageGetterAdapter.java
@@ -1,6 +1,6 @@
-package dev.arnoldatse.opensource.look4dev.core.fileStorage.adapters;
+package dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.adapters;
 
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.enums.FilesDirectories;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.enums.FilesDirectories;
 import dev.arnoldatse.opensource.look4dev.core.http.defaultExceptions.NotFoundException;
 
 public interface FileStorageGetterAdapter<T> {
