+++ core/src/main/java/dev/arnoldatse/opensource/look4dev/core/handleFiles/fileStorage/adapters/FileStoragePersistAndUrlGetterAdapter.java
@@ -1,4 +1,4 @@
-package dev.arnoldatse.opensource.look4dev.core.fileStorage.adapters;
+package dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.adapters;
 
 public interface FileStoragePersistAndUrlGetterAdapter<T> extends FileStoragePersistAdapter<T>, FileStorageUrlGetterAdapter{
 }
