+++ core/src/main/java/dev/arnoldatse/opensource/look4dev/core/handleFiles/fileStorage/adapters/FileStorageUrlGetterAdapter.java
@@ -1,6 +1,6 @@
-package dev.arnoldatse.opensource.look4dev.core.fileStorage.adapters;
+package dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.adapters;
 
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.enums.FilesTypesUrlsParts;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.enums.FilesTypesUrlsParts;
 import dev.arnoldatse.opensource.look4dev.core.http.defaultExceptions.NotFoundException;
 
 public interface FileStorageUrlGetterAdapter {
