+++ core/src/main/java/dev/arnoldatse/opensource/look4dev/core/users/usecases/RegisterUser.java
@@ -1,6 +1,6 @@
 package dev.arnoldatse.opensource.look4dev.core.users.usecases;
 
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.adapters.FileStorageUrlGetterAdapter;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.adapters.FileStorageUrlGetterAdapter;
 import dev.arnoldatse.opensource.look4dev.core.users.UserPasswordEncoder;
 import dev.arnoldatse.opensource.look4dev.core.entities.user.User;
 import dev.arnoldatse.opensource.look4dev.core.entities.user.dtos.UserResponseDto;
