+++ core/src/main/java/dev/arnoldatse/opensource/look4dev/core/users/usecases/userProfileDetails/profileCv/UpdateUserProfileCv.java
@@ -2,16 +2,15 @@ package dev.arnoldatse.opensource.look4dev.core.users.usecases.userProfileDetail
 
 import dev.arnoldatse.opensource.look4dev.core.entities.user.User;
 import dev.arnoldatse.opensource.look4dev.core.entities.user.dtos.userProfileDetailsDto.UserProfileDetailsFileUrlResponseDto;
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.FailedToStoreFileException;
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.adapters.FileStoragePersistAndUrlGetterAdapter;
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.enums.FilesDirectories;
-import dev.arnoldatse.opensource.look4dev.core.fileStorage.enums.FilesTypesUrlsParts;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.CheckSupportedFileExtension;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.FailedToStoreFileException;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.adapters.FileStoragePersistAndUrlGetterAdapter;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.enums.FilesDirectories;
+import dev.arnoldatse.opensource.look4dev.core.handleFiles.fileStorage.enums.FilesTypesUrlsParts;
 import dev.arnoldatse.opensource.look4dev.core.http.defaultExceptions.FileExtensionNotSupportedException;
 import dev.arnoldatse.opensource.look4dev.core.http.defaultExceptions.NotFoundException;
 import dev.arnoldatse.opensource.look4dev.core.users.UserRepository;
 import dev.arnoldatse.opensource.look4dev.core.utils.RandomString;
-
-import java.util.Arrays;
 
 public class UpdateUserProfileCv<T> {
     private final User user;
@@ -29,7 +28,7 @@ public class UpdateUserProfileCv<T> {
     }
 
     public UserProfileDetailsFileUrlResponseDto execute() throws FailedToStoreFileException, NotFoundException, FileExtensionNotSupportedException {
-        if (checkFileExtensionCompatibility()) {
+        if (CheckSupportedFileExtension.check(fileExtention, UserProfileCvAuthorizedExtensions.get())){
             deleteOldCv();
 
             String fileName = generateFileName();
@@ -41,11 +40,6 @@ public class UpdateUserProfileCv<T> {
         }
         throw new FileExtensionNotSupportedException();
 
-    }
-
-    private boolean checkFileExtensionCompatibility() {
-        return Arrays.stream(UserProfileCvAuthorizedExtensions.get())
-                .anyMatch(fileExtention::equalsIgnoreCase);
     }
 
     private void deleteOldCv() {
